% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asBranchComponent.R
\name{asBranchComponent}
\alias{asBranchComponent}
\title{Transform all data structures into BranchComponent}
\usage{
asBranchComponent(x)
}
\arguments{
\item{x}{data structure output by some clustering method 
(e.g. hclust, kmeans, dbscan, etc.)}
}
\value{
a matrix providing the mapping 
between data points and cluster id.
}
\description{
Transform all data structures into BranchComponent
}
\examples{
x <- kmeans(matrix(rnorm(100),nrow=50),centers=3)
asBranchComponent(x)
}
