% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTree.R
\name{combineClusterings}
\alias{combineClusterings}
\title{This is a new ensemble method for combining multiple clustering outcomes}
\usage{
combineClusterings(clustering1, clustering2, ...)
}
\arguments{
\item{clustering1}{clustering result of method 1}

\item{clustering2}{clustering result of method 2}

\item{...}{other clustering result}
}
\value{
final one hierarchical clustering
}
\description{
This is a new ensemble method for combining multiple clustering outcomes
}
\examples{
data <- rbind(matrix(rnorm(100, mean = 10, sd = 2), nrow = 50),
              matrix(rnorm(100, mean = 0, sd = 1), nrow = 50),
              matrix(rnorm(100, mean = -10, sd = 3), nrow = 50)
              )

clustering1<-stats::hclust(dist(data),method='single')
clustering2<-kmeans(data,centers=3)
clustering3<-dbscan::dbscan(data,eps=.1)
combineClusterings(clustering1,clustering2,clustering3)
}
