% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{createRectangleWithinRegion}
\alias{createRectangleWithinRegion}
\title{create a rectangle in a region with padding parameter}
\usage{
createRectangleWithinRegion(leftBoundary, bottomBoundary, rightBoundary,
  topBoundary, padding)
}
\arguments{
\item{leftBoundary}{left boundary of rectangle plot region}

\item{bottomBoundary}{bottom boundary of rectangle plot region}

\item{rightBoundary}{bottom boundary of rectangle plot region}

\item{topBoundary}{bottom boundary of rectangle plot region}

\item{padding}{padding of rectangle inside rectangle plot region}
}
\description{
create a rectangle in a region with padding parameter
}
