% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrend.R
\name{ctrend}
\alias{ctrend}
\title{Estimate cubic trend}
\usage{
ctrend(Y)
}
\arguments{
\item{Y}{an observation matrix. The row corresponds to time steps and the column corresponds to variables.}
}
\value{
trend: a matrix of fitted values.

dim: a matrix of dimensions of estimated trends.

coef: a matrix of estimates.
}
\description{
\code{ctrend} This function estimates cubic trends.
}
\examples{
#ctrend(Y)
}
