% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distToGram}
\alias{distToGram}
\title{transform clusDist object to Gram matrix}
\usage{
distToGram(x)
}
\arguments{
\item{x}{a clusDist object}
}
\description{
transform clusDist object to Gram matrix
}
\examples{
data <- rbind(matrix(rnorm(100, mean = 10, sd = 2), nrow = 50),
              matrix(rnorm(100, mean = 0, sd = 1), nrow = 50),
              matrix(rnorm(100, mean = -10, sd = 3), nrow = 50)
              )
clustering1 <- stats::hclust(dist(data),method='single')
clustering2 <- kmeans(data,centers=3)
clustering3 <- dbscan::dbscan(data,eps=.8)
distance <- clusDist(clustering1,clustering2,clustering3)
res <- distToGram(distance)
}
