% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getBranchInfo}
\alias{getBranchInfo}
\title{get plot information on a layer}
\usage{
getBranchInfo(x, layer = 1)
}
\arguments{
\item{x}{a clusterTreeDensityPlot object}
}
\value{
a list which contains plot information
}
\description{
get plot information on a layer
}
