% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getClusteringDistance}
\alias{getClusteringDistance}
\title{Calculate distance of two hierarchical clustering matrix}
\usage{
getClusteringDistance(branchComponent1, branchComponent2)
}
\arguments{
\item{branchComponent1}{The first clustering as a branch component}

\item{branchComponent2}{The second clustering as a branch component}
}
\value{
distance between 0 and square root of 2
}
\description{
Calculate distance of two hierarchical clustering matrix
}
\examples{
data <- rbind(matrix(rnorm(100, mean = 10, sd = 2), nrow = 50),
              matrix(rnorm(100, mean = 0, sd = 1), nrow = 50),
              matrix(rnorm(100, mean = -10, sd = 3), nrow = 50)
              )
clust1 <- asBranchComponent(stats::hclust(dist(data),method='complete'))
clust2 <- asBranchComponent(stats::hclust(dist(data),method='single'))
getClusteringDistance(clust1,clust2)
}
