% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getComponentsfromMerge}
\alias{getComponentsfromMerge}
\title{get the branch component family from the result of (single/avearge/complete) linkage clustering.}
\usage{
getComponentsfromMerge(merge)
}
\arguments{
\item{merge}{is the component from the result of "hclust", it is an n-1 by 2 matrix.}
}
\value{
a matrix which represent data id and cluster assignment
}
\description{
get the branch component family from the result of (single/avearge/complete) linkage clustering.
}
\examples{
data<-matrix(rnorm(100),nrow=50)
cl<-hclust(dist(data),method='single')
getComponentsfromMerge(cl$merge)
}
