% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTree.R
\name{groupIndexWithSameId}
\alias{groupIndexWithSameId}
\title{group index with same id in first column of treeMatrix}
\usage{
groupIndexWithSameId(treeMatrix, IDs, initialOrder)
}
\arguments{
\item{treeMatrix}{treeMatrix attribute of cluster tree}

\item{IDs}{cluster id in treeMatrix}

\item{initialOrder}{initialOrder is the initial order of treeMatrix}
}
\description{
group index with same id in first column of treeMatrix
}
