% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_asgn.R
\name{icon.asgn}
\alias{icon.asgn}
\title{Assign icons}
\usage{
icon.asgn(Z)
}
\arguments{
\item{Z}{an observation matrix. The row corresponds to time steps and the column corresponds to variables.}
}
\value{
a data.frame. This has the following items:
V which is names of variables, icon which is assigned icon, and fitted values of trends.
}
\description{
\code{icon.asgn} This function assigns icons to cubic trends.
}
\examples{
#icon.asgn(Z)
}
