% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_asgn.R
\name{icon.fit}
\alias{icon.fit}
\title{Calculate fitted values}
\usage{
icon.fit(x)
}
\arguments{
\item{x}{a vector of explanatory variables.}
}
\value{
a fitted value.
}
\description{
\code{icon.fit} This function calculates fitted values.
}
\examples{
#icon.fit(x)
}
