% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{nBranches}
\alias{nBranches}
\title{print number of branches of clusterTreePlotInfo object}
\usage{
nBranches(x)
}
\arguments{
\item{x}{a clusterTreePlotInfo object}
}
\value{
a integer, represents number of branches
}
\description{
print number of branches of clusterTreePlotInfo object
}
