% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{numBranches}
\alias{numBranches}
\title{print number of branches of clusterTreeDensityPlot object}
\usage{
numBranches(x)
}
\arguments{
\item{x}{a clusterTreeDensityPlot object}
}
\value{
a integer, represents number of branches
}
\description{
print number of branches of clusterTreeDensityPlot object
}
