% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTree.R
\name{plot.clusterTree}
\alias{plot.clusterTree}
\title{plot a clusterTree object}
\usage{
\method{plot}{clusterTree}(x, y = NULL, plot.labels = FALSE,
  labels = NULL, axes = TRUE, frame.plot = FALSE, ann = TRUE,
  main = "Cluster Tree Density Plot", sub = NULL, xlab = NULL,
  ylab = "Height", col = NULL, ...)
}
\arguments{
\item{x}{a clusterTree object}

\item{y}{NULL.  Will be ignored with a warning if non-NULL}

\item{plot.labels}{whether to plot labels}

\item{labels}{labels for each data point}

\item{axes}{whether to plot axis on the left}

\item{frame.plot}{whether to plot frame for density plot}

\item{ann}{whether to annotate main, sub, xlab, ylab}

\item{main}{main title for the plot}

\item{sub}{subtitle for the plot}

\item{xlab}{label for x-axis}

\item{ylab}{label for y-axis}

\item{col}{color of rectangle}

\item{...}{remains to be processed}
}
\value{
No return value (invisible())
}
\description{
plot a clusterTree object
}
\examples{
data <- rbind(matrix(rnorm(100, mean = 10, sd = 2), nrow = 50),
              matrix(rnorm(100, mean = 0, sd = 1), nrow = 50),
              matrix(rnorm(100, mean = -10, sd = 3), nrow = 50)
              )
clustering1<-stats::hclust(dist(data),method='single')
clustering2<-kmeans(data,centers=3)
clustering3<-dbscan::dbscan(data,eps=.1)
res <- combineClusterings(clustering1,clustering2,clustering3)
plot(res)
}
