% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotClusterTreeDensityPlotHelper}
\alias{plotClusterTreeDensityPlotHelper}
\title{helper function to plot a clusterTreeDensityPlot object}
\usage{
plotClusterTreeDensityPlotHelper(x, labels.plot, labels, col, labels.cex,
  labels.col)
}
\description{
helper function to plot a clusterTreeDensityPlot object
}
