% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotClusterTreeHelper}
\alias{plotClusterTreeHelper}
\title{a helper function of plot.clusterTree
this function plot clusterTree object recursively}
\usage{
plotClusterTreeHelper(x, xleft, ybottom, xright, ytop, labels.plot, labels,
  col, labels.cex = 1, labels.col)
}
\arguments{
\item{x}{a matrix which is tree attribute of clusterTree object}

\item{xleft}{left x coordinate of rectangular plot region}

\item{ybottom}{lower y coordinate of rectangular plot region}

\item{xright}{left x coordinate of rectangular plot region}

\item{ytop}{upper top coordinate of rectangular plot region}

\item{labels.plot}{whether to plot labels}

\item{labels}{labels for the plot}

\item{col}{color of rectangle}

\item{labels.cex}{a character expansion multiplier for the size of the object labels in the plot}

\item{labels.col}{colours of the object labels in the plot}
}
\value{
Invisibly returns the coordinates of the bottom level rectangles, each as
 (left bottom right top)
}
\description{
a helper function of plot.clusterTree
this function plot clusterTree object recursively
}
