% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotClusterTreePlotInfoRecursiveHelper}
\alias{plotClusterTreePlotInfoRecursiveHelper}
\title{helper function to plot a clusterTreePlotInfo object}
\usage{
plotClusterTreePlotInfoRecursiveHelper(x, labels.plot, labels, col,
  labels.cex, labels.col, ...)
}
\arguments{
\item{x}{a clusterTreePlotInfo object}

\item{labels.plot}{whether to plot labels}

\item{labels}{labels of data points}

\item{col}{color of rectangles}

\item{labels.cex}{cex of labels}

\item{labels.col}{color of labels}

\item{...}{other parameters to pass to graphics::rect()}
}
\description{
helper function to plot a clusterTreePlotInfo object
}
