% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{print.clusDist}
\alias{print.clusDist}
\title{print distance(a clusDist object) between clustering results}
\usage{
\method{print}{clusDist}(x, diag = NULL, upper = NULL,
  digits = getOption("digits"), justify = "none", right = TRUE, ...)
}
\arguments{
\item{x}{a clusDist object}

\item{diag}{whether to plot diagonal of distance matrix}

\item{upper}{whether to plot upper half of distance matrix}

\item{digits}{precision digits to retain}

\item{justify}{should a character vector be left-justified (the default), right-justified, centred or left alone. Can be abbreviated.}

\item{right}{logical, indicating whether or not strings should be right aligned.}

\item{...}{remain to be processed}
}
\description{
print distance(a clusDist object) between clustering results
}
