% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{showLabels}
\alias{showLabels}
\title{show coordinates of labels in clusterTreePlotInfo object}
\usage{
showLabels(x)
}
\arguments{
\item{x}{a clusterTreePlotInfo object}
}
\value{
a list of labels' coordinates
}
\description{
show coordinates of labels in clusterTreePlotInfo object
}
