% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{showLines}
\alias{showLines}
\title{show coordinates of lines in clusterTreePlotInfo object}
\usage{
showLines(x)
}
\arguments{
\item{x}{a clusterTreePlotInfo object}
}
\value{
a list of lines' coordinates
}
\description{
show coordinates of lines in clusterTreePlotInfo object
}
