% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{showRectangles}
\alias{showRectangles}
\title{show coordinates of rectangles in clusterTreeDensityPlot object}
\usage{
showRectangles(x)
}
\arguments{
\item{x}{a clusterTreeDensityPlot object}
}
\value{
a list of rectangle coordinates
}
\description{
show coordinates of rectangles in clusterTreeDensityPlot object
}
