% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testtrec.R
\name{testtrec}
\alias{testtrec}
\title{test}
\usage{
testtrec(v)
}
\arguments{
\item{v}{a number}
}
\value{
some figures for trends and assigned icons
}
\description{
\code{testtrec} This is test function.
}
\examples{
#testtrec(v)
}
