% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREC2_1.R
\name{TREC2.1}
\alias{TREC2.1}
\title{The additional step of the second step of TREC}
\usage{
TREC2.1(trn, argTREC, groups = 2)
}
\arguments{
\item{trn}{a list. This has thee elements of which each element is a vector of indexes of variables for increasing/flat/decreasing trends.}

\item{argTREC}{the output "argTREC" of TREC1}

\item{groups}{The number of groups corresponding to "trn".}
}
\value{
a figure
}
\description{
\code{TREC2.1} This function draws trends for each group.
}
\examples{
#TREC2.1(trn, argTREC)
}
