% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREC3.R
\name{TREC3}
\alias{TREC3}
\title{The third step of TREC}
\usage{
TREC3(tvar, trn, argTREC)
}
\arguments{
\item{tvar}{a vector of variable names for target trends.}

\item{trn}{the output \code{trn} of \code{TREC2}.}

\item{argTREC}{the output \code{argTREC} of \code{TREC1}.}
}
\value{
some figures for trends and assigned icons
}
\description{
\code{TREC3} This function performs clustering for trends.
}
\examples{
#TREC3(tvar, trn, argTREC)
}
