% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{TRECgetComponentsfromClustering}
\alias{TRECgetComponentsfromClustering}
\title{This is a new ensemble method for combining multiple clustering outcomes}
\usage{
TRECgetComponentsfromClustering(clustering)
}
\arguments{
\item{clustering:}{a matrix which stands for multiple clustering outcomes each row represent a data point, each column is an assignment of cluster id}
}
\value{
final one hierarchical clustering
}
\description{
This is a new ensemble method for combining multiple clustering outcomes
}
\examples{
data<-matrix(rnorm(100),nrow=50)
cl1<-asBranchComponent(hclust(dist(data),method='single'))
cl2<-asBranchComponent(kmeans(data,centers=3))
TRECgetComponentsfromClustering(cbind(cl1,cl2))
}
