% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asBranchComponent.R
\name{asBranchComponent}
\alias{asBranchComponent}
\title{Transform all data structures into BranchComponent}
\usage{
asBranchComponent(x)
}
\arguments{
\item{input}{data structure}
}
\value{
a matrix
}
\description{
Transform all data structures into BranchComponent
}
\examples{
x<-kmeans(matrix(rnorm(100),nrows=50),centers=3)
asBranchComponent(x)
}
