% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getClusteringDistance}
\alias{getClusteringDistance}
\title{calculate distance of two hierarchical clustering matrix}
\usage{
getClusteringDistance(clustering1, clustering2)
}
\arguments{
\item{two}{hierarchical clustering matrix}
}
\value{
distance between 0 and square root of 2
}
\description{
calculate distance of two hierarchical clustering matrix
}
\examples{
data<-matrix(rnorm(100),nrow=50)
clu1<-asBranchComponent(hclust(dist(data),method='complete'))
clu2<-asBranchComponent(hclust(dist(data),method='single'))
getClusteringDistance(clu1,clu2)
}
