% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getComponentsFromMerge}
\alias{getComponentsFromMerge}
\title{Get the branch component family from the result of (single/average/complete) linkage clustering.}
\usage{
getComponentsFromMerge(merge)
}
\arguments{
\item{merge}{The n-1 by 2 matrix that is the merge component output from
a hierarchical clustering which describes how the n points are merged in the cluster
hierarchy.}
}
\value{
An n by nlevels matrix where each row corresponds to a data point and each
column identifies a cluster number for that data point in the hierarchy.
}
\description{
Get the branch component family from the result of (single/average/complete) linkage clustering.
}
\examples{
data <- rbind(matrix(rnorm(100, mean = 10, sd = 2), nrow = 50),
              matrix(rnorm(100, mean = 0, sd = 1), nrow = 50),
              matrix(rnorm(100, mean = -10, sd = 3), nrow = 50)
              )
              
cl <- stats::hclust(dist(data),method='single')
getComponentsFromMerge(cl$merge)

}
