% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss_inpol.R
\name{miss.inpol}
\alias{miss.inpol}
\title{Missing interpolation}
\usage{
miss.inpol(X, remove.num)
}
\arguments{
\item{X}{a matrix of an observed data.}

\item{remove.num}{a variable, which has missing values of which the number is larger than \code{remove.num}, is removed.}
}
\value{
a matrix of the observed data with a linear interpolation.
}
\description{
\code{miss.inpol} This function takes a linear interpolation to missing data.
}
\examples{
#miss.inpol(X)
}
