% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTree.R
\name{plot.clusterTree}
\alias{plot.clusterTree}
\title{plot a clusterTree object}
\usage{
\method{plot}{clusterTree}(x, y = NULL, ...)
}
\arguments{
\item{x}{a clusterTree object}
}
\value{
...remains to be processed
}
\description{
plot a clusterTree object
}
\examples{
data <- rbind(matrix(rnorm(100, mean = 10, sd = 2), nrow = 50),
              matrix(rnorm(100, mean = 0, sd = 1), nrow = 50),
              matrix(rnorm(100, mean = -10, sd = 3), nrow = 50)
              )
clustering1<-stats::hclust(dist(data),method='single')
clustering2<-kmeans(data,centers=3)
clustering3<-dbscan::dbscan(data,eps=.1)
res <- combineClusterings(clustering1,clustering2,clustering3)
plot(res)
}
