% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotClusterTreeDensityPlotHelper}
\alias{plotClusterTreeDensityPlotHelper}
\title{helper function to plot a clusterTreeDensityPlot object}
\usage{
plotClusterTreeDensityPlotHelper(x, labels.plot, labels, col, labels.cex,
  labels.col)
}
\arguments{
\item{x}{a clusterTreeDensityPlot object}

\item{labels.plot}{whether to plot labels}

\item{labels}{of data points}

\item{col}{color of rectangles}

\item{labels.cex}{cex of labels}

\item{labels.col}{color of labels}
}
\description{
helper function to plot a clusterTreeDensityPlot object
}
