% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTree.R
\name{plotClusterTreeHelper}
\alias{plotClusterTreeHelper}
\title{a helper function of plot.clusterTree}
\usage{
plotClusterTreeHelper(x, xleft, ybottom, xright, ytop)
}
\arguments{
\item{x}{a matrix which is tree attribute of clusterTree object}

\item{xleft}{left x coordinate of rectangular plot region}

\item{ybottom}{lower y coordinate of rectangular plot region}

\item{xright}{left x coordinate of rectangular plot region}

\item{ytop}{upper top coordinate of rectangular plot region}
}
\description{
a helper function of plot.clusterTree
}
