% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTree.R
\name{plotClusterTreeHelper}
\alias{plotClusterTreeHelper}
\title{a helper function of plot.clusterTree}
\usage{
plotClusterTreeHelper(x, xleft, ybottom, xright, ytop, plot.labels, labels,
  col)
}
\arguments{
\item{x}{a matrix which is tree attribute of clusterTree object}

\item{xleft}{left x coordinate of rectangular plot region}

\item{ybottom}{lower y coordinate of rectangular plot region}

\item{xright}{left x coordinate of rectangular plot region}

\item{ytop}{upper top coordinate of rectangular plot region}

\item{plot.labels}{whether to plot labels}

\item{labels}{labels for the plot}

\item{col}{color of rectangle}
}
\value{
Invisibly returns the coordinates of the bottom level rectangles, each as
 (left bottom right top)
}
\description{
a helper function of plot.clusterTree
}
