% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTree.R
\name{reOrderClusterTreeMatrix}
\alias{reOrderClusterTreeMatrix}
\title{reorder rows of tree attribute of clusterTree object}
\usage{
reOrderClusterTreeMatrix(x, labels = NULL)
}
\arguments{
\item{x}{is the tree attribute of clusterTree object}

\item{labels}{labels is the order of rows of x}
}
\value{
an order which simplies the process of plotting dendogram/density plot
}
\description{
reorder rows of tree attribute of clusterTree object
}
