% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{showLabels}
\alias{showLabels}
\title{show coordinates of labels in clusterTreeDensityPlot object}
\usage{
showLabels(x)
}
\arguments{
\item{x}{a clusterTreeDensityPlot object}
}
\value{
a list of labels' coordinates
}
\description{
show coordinates of labels in clusterTreeDensityPlot object
}
