% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{showTreeMatrix}
\alias{showTreeMatrix}
\title{show treeMatrix of clusterTreePlotInfo object}
\usage{
showTreeMatrix(x)
}
\arguments{
\item{x}{a clusterTreePlotInfo object}
}
\value{
a matrix, represents treeMatrix
}
\description{
show treeMatrix of clusterTreePlotInfo object
}
