% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREC1.R
\name{TREC1}
\alias{TREC1}
\title{The first step of TREC}
\usage{
TREC1(Y)
}
\arguments{
\item{Y}{an observation matrix. The row corresponds to time steps and the column corresponds to variables.}
}
\value{
some figures for observed data and estimated trend, and arguments for next steps.
}
\description{
\code{TREC1} This function estimates cubic trends for multivariate time series observations.
}
\examples{
#TREC1(Y)
}
