% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREC1.R
\name{TREC1}
\alias{TREC1}
\title{The first step of TREC}
\usage{
TREC1(Y, time.points = NULL, remove.num = NULL)
}
\arguments{
\item{Y}{an observation matrix or data.frame; the row corresponds to time steps and the column corresponds to variables.}

\item{time.points}{a column name of \code{Y} expressing time points, if need; default is \code{NULL}.}

\item{remove.num}{a variable, which has missing values of which the number is larger than \code{remove.num}, is removed.
Default is the half of the number of time points.}
}
\value{
a list of results for trend estimation which has the following elements:
\item{fig.RawData}{a figure of raw data plot;
if the number of variables is larger than 16, this is a list of figures}

\item{fig.StdData}{a figure of standardized data plot;
if the number of variables is larger than 16, this is a list of figures}

\item{fig.ctrend}{a figure of estimated trends with prediction band for each variable;
if the number of variables is larger than 16, this is a list of figures}

\item{fig.trend}{a figure of all estimated trends}

\item{argTREC}{a list required in \code{TREC2}}

\item{remove}{removed variable names}

\item{Vnames}{a relationship between the original variable names and
the represented variable names}
\item{empty}{empty variable names}
}
\description{
\code{TREC1} This function estimates cubic trends for multivariate time series observations.
}
\examples{
#TREC1(Y)
}
