% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREC2.R
\name{TREC2}
\alias{TREC2}
\title{The second step of TREC}
\usage{
TREC2(argTREC, pvar = NULL, groups = 2)
}
\arguments{
\item{argTREC}{the output "argTREC" of TREC1}

\item{pvar}{two variable names for representative trends (option)}

\item{groups}{the number of groups for classification}
}
\value{
a dendrogram
}
\description{
\code{TREC2} This function performs clustering for trends.
}
\examples{
#TREC2(argTREC)
}
