% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREC2.R
\name{TREC2}
\alias{TREC2}
\title{The second step of TREC}
\usage{
TREC2(pnum, argTREC)
}
\arguments{
\item{pnum}{a vector of two indexes of variables for representative trends}

\item{argTREC}{the output "argTREC" of TREC1}
}
\value{
a dendrogram
}
\description{
\code{TREC2} This function performs clustering for trends.
}
\examples{
#TREC2(pnum, argTREC)
}
