% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREC2.R
\name{TREC2}
\alias{TREC2}
\title{The second step of TREC (v1.1.0)}
\usage{
TREC2(argTREC, clustering = TRUE, pvar = NULL, groups = 2)
}
\arguments{
\item{argTREC}{the output \code{argTREC} of \code{TREC1}}

\item{clustering}{if \code{TRUE}, clustering by a dendorogram (default);
if \code{FALSE}, discrimination by the criterion}

\item{pvar}{two variable names for representative trends (option)}

\item{groups}{the number of groups for classification}
}
\value{
a list of classification results which has the following elements:
\item{dend}{a fanction with no parameter to draw a dendrogram}

\item{fig.trends}{a figure of estimated trends for each classified group}

\item{trn}{a list of variable names for each classified group}
}
\description{
\code{TREC2} This function performs rough classification of trends.
}
\examples{
#TREC2(argTREC)
}
