% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREC3.R
\name{TREC3}
\alias{TREC3}
\title{The third step of TREC}
\usage{
TREC3(tvar, trn, argTREC)
}
\arguments{
\item{tvar}{a vector of variable names for target trends.}

\item{trn}{the output \code{trn} of \code{TREC2}.}

\item{argTREC}{the output \code{argTREC} of \code{TREC1}.}
}
\value{
a list of concrete classification results which has the following elements:
\item{fig.down}{a figure of estimated trends for each group in "Downward";
if the number of variables of downward group is larger than 16, this is a list of figures}

\item{fig.up}{a figure of estimated trends for each group in "Upward";
if the number of variables of upward group is larger than 16, this is a list of figures}

\item{fig.flat}{a figure of estimated trends for each group in "Flat";
if the number of variables of flat group is larger than 16, this is a list of figures}

\item{fig.icon}{summary}
}
\description{
\code{TREC3} This function performs clustering for trends.
}
\examples{
#TREC3(tvar, trn, argTREC)
}
