% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREC3.R
\name{TREC3}
\alias{TREC3}
\title{The third step of TREC}
\usage{
TREC3(tvar, argTREC)
}
\arguments{
\item{tvar}{a vector of variable labels for target trends.}

\item{argTREC}{the output "argTREC" of TREC1.}
}
\value{
some figures for trends and assigned icons
}
\description{
\code{TREC3} This function performs clustering for trends.
}
\examples{
#TREC3(tnum, argTREC)
}
